1.4.0 release.
Squashed commit of the following:

commit 4274586a77465154c997ee823cdcdd074d183876
Author: Jagadeesh Kotra <jagadeesh@stdin.top>
Date:   Mon Apr 4 23:17:59 2022 +0530

    bugfix: -d should be expanded iff `.`

commit bcd2df0405c17028755ecff0b973ffec7f63c36d
Author: Jagadeesh Kotra <jagadeesh@stdin.top>
Date:   Mon Apr 4 19:40:49 2022 +0530

    Expand . to current dir

    * if -d / --directory arg is just ., replace it with current dir.
    * wait 1 sec on every iteration of initial candidates.

commit 18c2465413562379e11207c581c1b8bb69e0d3bf
Author: Jagadeesh Kotra <jagadeesh@stdin.top>
Date:   Mon Apr 4 00:06:50 2022 +0530

    Improve json read/write in wallhaven

    * use `Serialize` and `Deserialize` of `serde` on `WallHaven` struct
for
    auto reading and writing to JSON file.

commit 995d8885c15668624d670a6be404f21af1c04566
Author: Jagadeesh Kotra <jagadeesh@stdin.top>
Date:   Mon Apr 4 00:04:27 2022 +0530

    Check if file already exists and return Error message instead of
panic.

commit 318f45173de2435e413dd58171ac367a66a87e27
Author: Jagadeesh Kotra <jagadeesh@stdin.top>
Date:   Thu Mar 31 00:57:06 2022 +0530

    * remove deprecated args filtering from misc.rs get_wpc_args()

    * use PathBuf .extension() to get for chosen wallpaper

commit 2b2ddb8b81cf9c566ac0f239a58f8690488fdde9
Author: Jagadeesh Kotra <jagadeesh@stdin.top>
Date:   Wed Mar 30 19:36:21 2022 +0530

    update CHANGELOG.md

commit 47ce46068c3ae92d2c1d4148ee08d95cf32a1197
Author: Jagadeesh Kotra <jagadeesh@stdin.top>
Date:   Wed Mar 30 19:34:15 2022 +0530

    enable lto on release build

commit df2cdbad6d7f41db878314e4572ce7db2d0835c5
Author: Jagadeesh Kotra <jagadeesh@stdin.top>
Date:   Wed Mar 30 18:46:44 2022 +0530

    refactor and simplify - v1.4.0

+++ src/misc.rs
@@ -4,7 +4,7 @@ use std::path::PathBuf;
 extern crate rand;
 use rand::Rng;
 
-use log::{debug, info, warn, error};
+use log::{debug, error};
 
 use std::env::current_exe;
 
@@ -60,16 +60,11 @@ pub fn get_wpc_args() -> Vec<String> {
     for i in (0..args.len()).rev(){
         if args[i] == "--startup"{ args.remove(i); }
         else if args[i] == "-S" { args.remove(i); }
-        else if args[i] == "--debug" { args.remove(i); }
-        else if args[i] == "-D" { args.remove(i); }
         else if args[i] == "--background" { args.remove(i); }
-        else if args[i] == "-d" || args[i] == "--directory"{
-            if args[i+1] == "."{
-
-                args[i+1] = String::from(std::env::current_dir().unwrap().to_str().unwrap())
-
+        else if args[i] == "-d" || args[i] == "--directory" {
+            if args[i+1] == "." {
+                args[i+1] = String::from(std::env::current_dir().unwrap().to_str().unwrap());
             }
-
         }
     }
 
@@ -108,16 +103,7 @@ pub async fn download_wallpapers(urls: Vec<String>, savepath: &str) -> Vec<Strin
             let mut filename = PathBuf::from(savepath);
             filename = filename.join(file_vec[file_vec.len() - 1]);
 
-            if url.contains("bing.com"){
-                filename.pop();
-                filename = filename.join("bing_wpod.jpeg");
-            }
-
             remote_files.push(String::from(filename.to_str().unwrap()));
-
-            if filename.exists() && !url.contains("bing.com"){
-                    continue
-            }
 
             match async_download(url.as_str(), filename.to_str().unwrap()).await{
                 Ok(_) => (),
@@ -132,30 +118,22 @@ pub async fn download_wallpapers(urls: Vec<String>, savepath: &str) -> Vec<Strin
 async fn async_download(url: &str, filename: &str) -> Result<bool, String> {
 
     let filedest = PathBuf::from(filename);
+    if filedest.exists() { return  Ok(true); }
     let response = match reqwest::get(url).await{
         Ok(f) => f,
-        Err(why) => panic!("WPC panic: {}", why)
+        Err(why) => return Err(String::from(format!("{:?}", why)))
     };
 
     let mut out = File::create(filedest).expect("failed to create file");
     let content = match response.bytes().await{
         Ok(f) => f,
-        Err(why) => panic!("WPC panic: {}", why)
+        Err(why) => return Err(String::from(format!("{:?}", why)))
     };
 
     let mut content = std::io::Cursor::new(content);
     io::copy(&mut content, &mut out).expect("failed to copy content");
 
     return Ok(true)
-}
-
-#[allow(dead_code)]
-pub fn download(url: &str, filename: &str) -> Result<(), Box<dyn std::error::Error>> {
-
-    let mut response = reqwest::blocking::get(url).expect("Cannot download!");
-    let mut out = File::create(filename).expect("failed to create file");
-    io::copy(&mut response, &mut out).expect("failed to copy content");
-    Ok(())
 }
 
 pub fn random_n(len_max: usize) -> usize {
@@ -190,27 +168,6 @@ pub fn update_file_list(dirpath: &str, maxage: i64) -> Vec<String> {
     return wallpapers
 }
 
-pub fn clean_gs(dirpath: &str) {
-
-    let mut gs_dir = PathBuf::from(dirpath);
-    gs_dir.push("grayscale");
-
-    if !gs_dir.exists(){
-        debug!("grayscale dir does not exist!");
-        return
-    }
-
-    let files = gs_dir.read_dir().unwrap();
-
-    for file in files {
-        let fp = file.unwrap().path().to_str().unwrap().to_string();
-
-        if fp.contains("_grayscale."){
-            let _ = std::fs::remove_file(fp);
-        }
-    }
-}
-
 pub fn maxage_filter(file_list: Vec<String>, maxage: i64) -> Vec<String>{
 
     if maxage == -1 { return file_list }
@@ -242,15 +199,15 @@ pub fn maxage_filter(file_list: Vec<String>, maxage: i64) -> Vec<String>{
 pub fn is_linux_gnome_de() -> bool {
     let res = std::env::var("DESKTOP_SESSION").unwrap().to_string();
     debug!("DESKTOP_SESSION = {}", res);
-    if res == "gnome".to_string() { return true }
-    if res == "gnome-xorg".to_string() { return true } //fedora
-    if res == "budgie-desktop".to_string() { return true } //budgie
+    if res.contains("gnome"){
+        return true;
+    }
     return false;
 }
 
 
 #[cfg(test)]
-mod bing {
+mod misc_tests {
 
     #[tokio::test]
     async fn async_download_test() {
